Changed the cast rather than the const

+++ APRSWriter.cpp
@@ -151,7 +151,7 @@ void CAPRSWriter::sendIdFrames()
 	if (m_txFrequency != 0U) {
 		float offset = float(int(m_rxFrequency) - int(m_txFrequency)) / 1000000.0F;
 		::sprintf(desc, "MMDVM Voice %.5LfMHz %c%.4lfMHz",
-			float(m_txFrequency) / 1000000.0L,
+			(long double)(m_txFrequency) / 1000000.0F,
 			offset < 0.0F ? '-' : '+',
 			::fabs(offset));
 	} else {
